# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MCQ Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/exams:
    get:
      operationId: adminListExams
      tags: [admin]
      summary: List all admin-created exams
      responses:
        "200":
          description: List of admin exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminExam"
    post:
      operationId: adminCreateExam
      tags: [admin]
      summary: Create a new admin exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminExamInput"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExam"
  /admin/exams/{id}:
    get:
      operationId: adminGetExam
      tags: [admin]
      summary: Get admin exam with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin exam details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExamDetail"
    patch:
      operationId: adminUpdateExam
      tags: [admin]
      summary: Update an admin exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminExamUpdate"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExam"
    delete:
      operationId: adminDeleteExam
      tags: [admin]
      summary: Delete an admin exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/exams/{id}/mcqs:
    get:
      operationId: adminGetExamMcqs
      tags: [admin]
      summary: Get MCQs for an admin exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: MCQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mcq"
...
  /admin-exams/{id}/start:
    post:
      operationId: startAdminExam
      tags: [admin]
      summary: Start an admin exam as a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Exam session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
...
    User:
      type: object
      required: [id, name, email, role, isPremium, isBanned, streakDays, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        isPremium:
          type: boolean
        isBanned:
          type: boolean
        streakDays:
          type: integer
        totalPoints:
          type: integer
        createdAt:
          type: string
...
    Mcq:
      type: object
      required: [id, chapterId, subjectId, question, options, correctOption, difficulty, isBookmarked, isWrong]
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        subjectId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: ["string", "null"]
        explanationImage:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        isBookmarked:
          type: boolean
        isWrong:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    ExamSession:
      type: object
      required: [id, userId, type, status, mcqs, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        status:
          type: string
          enum: [active, completed, expired]
        timeLimitMinutes:
          type: ["integer", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        mcqs:
          type: array
          items:
            $ref: "#/components/schemas/Mcq"
        createdAt:
          type: string
...
    UserListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin]
        isPremium:
          type: boolean
        isBanned:
          type: boolean
    BanInput:
      type: object
      required: [banned]
      properties:
        banned:
          type: boolean
...
    AdminExam:
      type: object
      required: [id, title, durationMinutes, totalMarks, passMarks, isPublished, createdBy, createdAt, questionCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        subjectId:
          type: integer
          nullable: true
        subjectName:
          type: string
          nullable: true
        chapterIds:
          type: array
          items:
            type: integer
        mcqIds:
          type: array
          items:
            type: integer
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        passMarks:
          type: integer
        isPublished:
          type: boolean
        questionCount:
          type: integer
        createdBy:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    AdminExamDetail:
      allOf:
        - $ref: "#/components/schemas/AdminExam"
        - type: object
          properties:
            mcqs:
              type: array
              items:
                $ref: "#/components/schemas/Mcq"
    AdminExamInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        chapterIds:
          type: array
          items:
            type: integer
        mcqIds:
          type: array
          items:
            type: integer
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        passMarks:
          type: integer
        isPublished:
          type: boolean
    AdminExamUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: integer
        chapterIds:
          type: array
          items:
            type: integer
        mcqIds:
          type: array
          items:
            type: integer
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        passMarks:
          type: integer
        isPublished:
          type: boolean
...