# lib/api-zod/src/generated/types/adminExam.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MCQ Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminExam {
  id: number;
  title: string;
  description?: string | null;
  subjectId?: number | null;
  subjectName?: string | null;
  chapterIds?: number[];
  mcqIds?: number[];
  durationMinutes: number;
  totalMarks: number;
  passMarks: number;
  isPublished: boolean;
  questionCount: number;
  createdBy: number;
  createdAt: string;
  updatedAt?: string;
}