# lib/api-zod/src/generated/types/adminExamDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MCQ Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminExam } from './adminExam';
import type { Mcq } from './mcq';

export type AdminExamDetail = AdminExam & {
  mcqs?: Mcq[];
};