# lib/api-zod/src/generated/types/adminExamInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MCQ Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminExamInput {
  title: string;
  description?: string;
  subjectId?: number;
  chapterIds?: number[];
  mcqIds?: number[];
  durationMinutes?: number;
  totalMarks?: number;
  passMarks?: number;
  isPublished?: boolean;
}